move models out to dedicated file
